split enums into common, conversation auxils and states
