+++ samanthas_telegram_bot/conversation/auxil/callback_query_reply_sender.py
@@ -6,6 +6,10 @@ from typing import Union
 from telegram import CallbackQuery, InlineKeyboardButton, InlineKeyboardMarkup
 from telegram.constants import ParseMode
 
+from samanthas_telegram_bot.conversation.auxil.enums import (
+    CommonCallbackData,
+    UserDataReviewCategory,
+)
 from samanthas_telegram_bot.data_structures.constants import (
     NON_TEACHING_HELP_TYPES,
     STUDENT_COMMUNICATION_LANGUAGE_CODES,
@@ -13,12 +17,7 @@ from samanthas_telegram_bot.data_structures.constants import (
     Locale,
 )
 from samanthas_telegram_bot.data_structures.context_types import CUSTOM_CONTEXT_TYPES
-from samanthas_telegram_bot.data_structures.enums import (
-    AgeRangeType,
-    CommonCallbackData,
-    Role,
-    UserDataReviewCategory,
-)
+from samanthas_telegram_bot.data_structures.enums import AgeRangeType, Role
 from samanthas_telegram_bot.data_structures.models import AssessmentQuestion
 
 logger = logging.getLogger(__name__)
