+++ samanthas_telegram_bot/conversation/callbacks/registration/student.py
@@ -8,12 +8,13 @@ from samanthas_telegram_bot.api_queries.smalltalk import send_user_data_get_smal
 from samanthas_telegram_bot.conversation.auxil.callback_query_reply_sender import (
     CallbackQueryReplySender as CQReplySender,
 )
+from samanthas_telegram_bot.conversation.auxil.enums import CommonCallbackData
 from samanthas_telegram_bot.conversation.auxil.message_sender import MessageSender
 from samanthas_telegram_bot.conversation.auxil.prepare_assessment import prepare_assessment
 from samanthas_telegram_bot.conversation.auxil.shortcuts import answer_callback_query_and_get_data
+from samanthas_telegram_bot.conversation.states.common import ConversationState
 from samanthas_telegram_bot.data_structures.constants import LEVELS_ELIGIBLE_FOR_ORAL_TEST, Locale
 from samanthas_telegram_bot.data_structures.context_types import CUSTOM_CONTEXT_TYPES
-from samanthas_telegram_bot.data_structures.enums import CommonCallbackData, ConversationState
 from samanthas_telegram_bot.data_structures.models import AssessmentAnswer
 
 logger = logging.getLogger(__name__)
