+++ samanthas_telegram_bot/conversation/states/common.py
@@ -1,0 +1,42 @@
+from enum import IntEnum, auto
+
+
+class ConversationState(IntEnum):
+    """Provides integer keys for the dictionary of states for ConversationHandler."""
+
+    IS_REGISTERED = auto()
+    CHECK_CHAT_ID_ASK_TIMEZONE = auto()
+    CHECK_IF_WANTS_TO_REGISTER_ANOTHER_PERSON_ASK_TIMEZONE = auto()
+    ASK_FIRST_NAME = auto()
+    ASK_LAST_NAME = auto()
+    ASK_SOURCE = auto()
+    CHECK_USERNAME = auto()
+    ASK_PHONE_NUMBER = auto()
+    ASK_EMAIL = auto()
+    ASK_ROLE = auto()
+    ASK_AGE = auto()
+    TIME_SLOTS_START = auto()
+    TIME_SLOTS_MENU_OR_ASK_TEACHING_LANGUAGE = auto()
+    ASK_LEVEL_OR_ANOTHER_TEACHING_LANGUAGE_OR_COMMUNICATION_LANGUAGE = auto()
+    ASK_LEVEL_OR_COMMUNICATION_LANGUAGE = auto()
+    ASK_TEACHING_EXPERIENCE = auto()
+    ASK_TEACHING_GROUP_OR_SPEAKING_CLUB = auto()
+    ADOLESCENTS_ASK_COMMUNICATION_LANGUAGE_OR_START_ASSESSMENT = auto()
+    ASK_STUDENT_NON_TEACHING_HELP_OR_START_REVIEW = auto()
+    ASK_ASSESSMENT_QUESTION = auto()
+    SEND_SMALLTALK_URL_OR_ASK_COMMUNICATION_LANGUAGE = auto()
+    ASK_COMMUNICATION_LANGUAGE_AFTER_SMALLTALK = auto()
+    ASK_NUMBER_OF_GROUPS_OR_TEACHING_FREQUENCY_OR_NON_TEACHING_HELP = auto()
+    ASK_TEACHING_FREQUENCY = auto()
+    PREFERRED_STUDENT_AGE_GROUPS_START = auto()
+    PREFERRED_STUDENT_AGE_GROUPS_MENU_OR_ASK_NON_TEACHING_HELP = auto()
+    NON_TEACHING_HELP_MENU_OR_PEER_HELP_FOR_TEACHER_OR_REVIEW_FOR_STUDENT = auto()
+    PEER_HELP_MENU_OR_ASK_ADDITIONAL_HELP = auto()
+    ASK_YOUNG_TEACHER_COMMUNICATION_LANGUAGE = auto()
+    ASK_YOUNG_TEACHER_SPEAKING_CLUB_LANGUAGE = auto()
+    ASK_YOUNG_TEACHER_ADDITIONAL_HELP = auto()
+    ASK_REVIEW = auto()
+    REVIEW_MENU_OR_ASK_FINAL_COMMENT = auto()
+    REVIEW_REQUESTED_ITEM = auto()
+    ASK_FINAL_COMMENT = auto()  # standalone, not after review
+    BYE = auto()
